requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
APScheduler==3.11.2
attrs==25.4.0
azure-core==1.38.2
azure-identity==1.25.2
azure-storage-blob==12.28.0
backoff==2.2.1
bcrypt==5.0.0
boto3==1.34.34
botocore==1.34.162
build==1.4.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
colorama==0.4.6
crewai==1.9.3
croniter==6.0.0
cryptography==43.0.3
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.115.14
fastapi-sso==0.16.0
filelock==3.24.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.78.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
hyperframe==6.1.0
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
instructor==1.12.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.1.0
json5==0.10.0
json_repair==0.25.3
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
litellm==1.75.3
litellm-enterprise==0.1.19
litellm-proxy-extras==0.2.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.23.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msal==1.35.0
msal-extensions==1.3.1
multidict==6.7.1
nodeenv==1.10.0
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.2
polars==1.38.1
polars-runtime-32==1.38.1
portalocker==2.7.0
postgrest==2.28.0
posthog==5.4.0
pre_commit==4.5.1
propcache==0.4.1
protobuf==5.29.6
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
PyNaCl==1.6.2
pyparsing==3.3.2
pypdfium2==5.5.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.18
pytz==2025.2
pywin32==311
PyYAML==6.0.3
realtime==2.28.0
redis==7.2.0
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.30.0
rq==2.7.0
s3transfer==0.10.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.46.2
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.20.3
tomli==2.0.2
tomli_w==1.1.0
tqdm==4.67.3
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uv==0.9.30
uvicorn==0.41.0
virtualenv==20.39.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==13.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
